requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
openai==1.3.0
pillow==10.1.0